Move IHandler to Interface package
create local variables in NewFilePage for readability
Task #25
