+++ de.tu_bs.cs.isf.e4cf.parts.project_explorer/src/de/tu_bs/cs/isf/e4cf/parts/project_explorer/interfaces/IHandler.java
@@ -1,4 +1,4 @@
-package de.tu_bs.cs.isf.e4cf.parts.project_explorer.handlers;
+package de.tu_bs.cs.isf.e4cf.parts.project_explorer.interfaces;
 
 import org.eclipse.e4.core.contexts.IEclipseContext;
 import org.eclipse.swt.widgets.Shell;
@@ -6,7 +6,7 @@ import org.eclipse.swt.widgets.Shell;
 import de.tu_bs.cs.isf.e4cf.core.util.ServiceContainer;
 import de.tu_bs.cs.isf.e4cf.core.util.services.RCPSelectionService;
 
-interface IHandler {
+public interface IHandler {
 
 	void execute(IEclipseContext context, ServiceContainer services, Shell shell);
